Weapon

See merge request mod-team-proj-2018/aloneinthedark!20

+++ src/main/java/shared/gameObjects/weapons/Punch.java
@@ -21,12 +21,12 @@ public class Punch extends Melee {
    */
   public Punch(
       double x,
-      double y,
-      double sizeX,
-      double sizeY,
-      ObjectID id,
-      double damage,
-      String name,
+      double y, 
+      double sizeX, 
+      double sizeY, 
+      ObjectID id, 
+      int damage,
+      String name, 
       double range,
       UUID uuid) {
     super(x, y, sizeX, sizeY, id, damage, 1, name, -1, 60, range, 1, 1, uuid);
